Support for eGroupware 1.6 instead of 1.4
Added eGroupware 'bookmarks' application
eGroupware backups integrated with eBox backups (closes #1290)



git-svn-id: http://svn.zentyal.org/zentyal/trunk@13951 866bc4c1-54da-0310-a958-9d5a9a0e8d04

+++ client/egroupware/src/EBox/EGroupwareLdapUser.pm
@@ -349,6 +349,9 @@ sub _setTemplate # (template, account) account = (uid | gid) number
     my @ids = @{$model->enabledRows()};
     my @apps = map ($model->row($_)->valueByName('app'), @ids);
 
+    unless ($self->_databaseExists()) {
+        return;
+    }
     # Insert new permissions (changepassword permission only for user accounts)
     my $sql = $account > 0 ?
               'INSERT INTO egw_acl ' .
@@ -401,8 +404,19 @@ sub _deletePermissions # (account) account = (uid | gid) number
 {
     my ($self, $account) = @_;
 
-    my $sql = "DELETE FROM egw_acl WHERE acl_account=$account;";
-    EBox::Sudo::root("su postgres -c \"psql egroupware -c \\\"$sql\\\"\"");
+    if ($self->_databaseExists()) {
+        my $sql = "DELETE FROM egw_acl WHERE acl_account=$account;";
+        EBox::Sudo::root("su postgres -c \"psql egroupware -c \\\"$sql\\\"\"");
+    }
+}
+
+sub _databaseExists
+{
+    my ($self) = @_;
+
+    my $cmd = "su postgres -c \"psql egroupware -c ''\"";
+    EBox::Sudo::rootWithoutException($cmd);
+    return ($? == 0);
 }
